requirements.txt
langchain
openai
chromadb
unstructured
beautifulsoup4
python-magic
python-dotenv